// visual-programming-engine/src/module/packages/maxGraph/core/src/types.ts
export type ColorValue = string;
/** Color values and special placeholders used to resolve colors (see {@link CellRenderer.resolveColor}) for style properties. */
export type SpecialStyleColorValue =
  | 'indicated'
  | 'inherit'
  | 'none'
  | 'swimlane'
  | (string & {});

export type DirectionValue = 'north' | 'south' | 'east' | 'west';
export type TextDirectionValue = '' | 'ltr' | 'rtl' | 'auto';
export type AlignValue = 'left' | 'center' | 'right';
export type VAlignValue = 'top' | 'middle' | 'bottom';
export type OverflowValue = 'fill' | 'width' | 'auto' | 'hidden' | 'scroll' | 'visible';
// ...
export type CanvasState = {
  alpha: number;
  dashPattern: string;
  dashed: boolean;
  dx: number;
  dy: number;
  fillAlpha: number;
  fillColor: ColorValue;
  fixDash: boolean;
  fontBackgroundColor: ColorValue;
  fontBorderColor: ColorValue;
  fontColor: ColorValue;
  fontFamily: string;
  fontSize: number;
  /**
   * See {@link CellStateStyle.fontStyle}.
   */
  fontStyle: number;
  gradientAlpha: number;
  gradientColor: ColorValue;
  gradientDirection: DirectionValue;
  gradientFillAlpha: number;
  lineCap: string;
  lineJoin: string;
  miterLimit: number;
  rotation: number;
  rotationCx: number;
  rotationCy: number;
  scale: number;
  shadow: boolean;
  shadowAlpha: number;
  shadowColor: ColorValue;
  shadowDx: number;
  shadowDy: number;
  strokeAlpha: number;
  strokeColor: ColorValue;
  strokeWidth: number;
  transform: string | null;
};